fix(clippy): resolve all clippy warnings in PR #73
- Prefix unused variables with underscore (_path, _block_num, _ino)
- Remove unnecessary mut from elf_data array
- Remove unnecessary cast (u64 -> u64)
- Mark parse_modules as unsafe with Safety doc comment
- Add #[allow(dead_code)] to simple_hash_str (used in tests)

Model: claude-sonnet-4-6
Tool: claude-code
Prompt: Fix clippy warnings in busybox integration PR

+++ src/kernel/elf.rs
@@ -573,7 +573,7 @@ pub fn start_user_program(
     let argc = unsafe { *argc_ptr };
 
     // argv starts right after argc
-    let argv_ptr = (context.stack_ptr + 8) as u64;
+    let argv_ptr = context.stack_ptr + 8;
 
     crate::serial::write_str("[elf] argc=");
     print_hex_u64(argc);
